pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bansho"
version = "0.1.0"
description = "Security gateway for the Model Context Protocol (MCP)"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.31.0",
  "httpx>=0.28.0",
  "mcp>=1.26.0",
  "pydantic>=2",
  "pydantic-settings>=2.11.0",
  "pyyaml>=6.0.0",
  "redis>=6.4.0",
  "structlog>=25.4.0",
]

[project.scripts]
bansho = "bansho.main:main"

[dependency-groups]
dev = [
  "mypy>=1.19.0",
  "pytest>=9.0.0",
  "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
